adding update part1

+++ src/eg/edu/alexu/csd/oop/db/cs2/controller/DatabaseManager.java
@@ -7,12 +7,15 @@ import eg.edu.alexu.csd.oop.db.cs2.conditions.*;
 import eg.edu.alexu.csd.oop.db.cs2.structures.*;
 
 import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Spliterator;
 import java.util.regex.*;
 
 public class DatabaseManager implements Database {
     private static DatabaseManager instance = new DatabaseManager();
+    private List<DatabaseContainer> databases = new ArrayList<>();
     private DatabaseContainer currentDatabase;
     private FilesHandler filesHandler = new FilesHandler();
     private ConditionsFilter equal = new Equal();
@@ -55,7 +58,10 @@ public class DatabaseManager implements Database {
             Matcher match = regex.matcher(query);
             match.find();
             String databaseName = match.group().toLowerCase().replaceAll("\\s+", "").replaceAll(";", "");
+            if (filesHandler.isDatabaseExist(databaseName))
+                return false;
             currentDatabase = new DatabaseContainer(databaseName);
+            databases.add(currentDatabase);
             filesHandler.createDatabase(databaseName);
             return true;
         }else if(QueriesParser.checkDropDatabase(query)){
@@ -63,7 +69,19 @@ public class DatabaseManager implements Database {
             Matcher match = regex.matcher(query);
             match.find();
             String databaseName = match.group().toLowerCase().replaceAll("\\s+", "").replaceAll(";", "");
-            currentDatabase = null;
+            if(!filesHandler.isDatabaseExist(databaseName))
+                return false;
+            if (currentDatabase.getName().equalsIgnoreCase(databaseName)){
+                databases.remove(currentDatabase);
+                currentDatabase = (databases.size() < 1 ) ? null : databases.get(databases.size()-1);
+            }else{
+                for (int i = 0; i < databases.size(); ++i){
+                    if (databases.get(i).getName().equalsIgnoreCase(databaseName)) {
+                        databases.remove(i);
+                        break;
+                    }
+                }
+            }
             filesHandler.dropDatabase(databaseName);
             return true;
         } else if (QueriesParser.checkCreateTable(query)) {
@@ -86,17 +104,20 @@ public class DatabaseManager implements Database {
             throw new SQLException();
         }
     }
-
     @Override
     public Object[][] executeQuery(String query) throws SQLException {
         Object[][] objects;
         query = query.toLowerCase();
         if (!QueriesParser.checkExecuteQuery(query))
             throw new SQLException();
-        if (query.matches("^\\s*select\\s*\\*\\s*from\\s+\\w+\\s*(\\s+where\\s+\\w+\\s*[=<>]\\s*([0-9]+|(\\'|\\\")\\w+(\\'|\\\")))?\\s*;?\\s*$")){
+        if (query.matches("^\\s*select\\s*\\*\\s*from\\s+\\w+\\s*(\\s+where\\s+\\w+\\s*[=<>]\\s*([0-9]+|(\\'\\w+\\')))?\\s*;?\\s*$")){
             query = query.replaceAll("^\\s*select\\s*\\*\\s*from\\s+", "").replaceAll("\\s*;?\\s*$", "");
             if(query.matches("^\\w+\\s+where\\s+\\w+\\s*[=<>]\\s*([0-9]+|\\'\\w+\\')$")){
                 String[] split = parseQuery(query);
+                if(!filesHandler.isTableExist(split[0], currentDatabase.getName()))
+                    throw new SQLException();
+                if (!currentDatabase.getTable(split[0]).containColumn(split[2]))
+                    throw new SQLException();
                 Table table = aSwitch.meetCondition(split[3], currentDatabase.getTable(split[0]), split[2], split[4]);
                 List<Column> columns = table.getColumns();
                 if (flag) {
@@ -106,6 +127,8 @@ public class DatabaseManager implements Database {
                 objects = selectTable(table.getColumns());
             }else {
                 Table table = currentDatabase.getTable(query);
+                if(!filesHandler.isTableExist(table.getName(), currentDatabase.getName()))
+                    throw new SQLException();
                 List<Column> columns = table.getColumns();
                 if (flag) {
                     columns.remove(0);
@@ -122,9 +145,21 @@ public class DatabaseManager implements Database {
             query = query.replaceAll("^(\\w+\\s*,\\s*)*\\w+\\s+from\\s+", "");
             if (query.matches("\\w+")){
                 Table table = currentDatabase.getTable(query);
+                if(!filesHandler.isTableExist(table.getName(), currentDatabase.getName()))
+                    throw new SQLException();
+                for (String columnName : columns){
+                    if(!table.containColumn(columnName))
+                        throw new SQLException();
+                }
                 objects = selectTable(table.getColumns(columns));
             }else{
                 String[] split = parseQuery(query);
+                if(!filesHandler.isTableExist(split[0], currentDatabase.getName()))
+                    throw new SQLException();
+                for (String columnName : columns){
+                    if(!currentDatabase.getTable(split[0]).containColumn(columnName))
+                        throw new SQLException();
+                }
                 Table table = aSwitch.meetCondition(split[3], currentDatabase.getTable(split[0]), split[2], split[4]);
                 objects = selectTable(table.getColumns(columns));
             }
@@ -134,6 +169,8 @@ public class DatabaseManager implements Database {
 
     @Override
     public int executeUpdateQuery(String query) throws SQLException {
+        if(currentDatabase == null)
+            return 0;
         if (QueriesParser.checkInsertInto(query)){
             HashMap<String, String> hashMap = new HashMap<>();
             if (query.toLowerCase().matches("^\\s*insert\\s+into\\s+\\w+\\s*\\((\\s*\\w+\\s*,)*\\s*\\w+\\s*\\)\\s*values\\s*\\((\\s*([0-9]+|\\'\\w+\\')\\s*,)*\\s*([0-9]+|\\'\\w+\\')\\s*\\)\\s*;?\\s*$")){
@@ -164,93 +201,39 @@ public class DatabaseManager implements Database {
                 currentDatabase.insertRow(split[0].toLowerCase(), split);
             }
             return 1;
-        }else if (QueriesParser.checkDeleteFromTable(query)){
+        }
+        else if (QueriesParser.checkDeleteFromTable(query)){
             query = query.toLowerCase();
             query = query.replaceAll("^\\s*delete\\s+from\\s", "").replaceAll("\\s*;?\\s*$", "");
             if(query.matches("^\\w+$")){
+                if (!filesHandler.isTableExist(query, currentDatabase.getName()))
+                    throw new SQLException();
                 return currentDatabase.clearTable(query);
             }else{
                 String[] split = parseQuery(query);
+                if(!filesHandler.isTableExist(split[0], currentDatabase.getName()) || !(currentDatabase.getTable(split[0]).containColumn(split[2])))
+                    throw new SQLException();
                 Table table = aSwitch.meetCondition(split[3], currentDatabase.getTable(split[0]), split[2], split[4]);
                 return currentDatabase.deleteItems(split[0], table);
             }
         }
         else if(QueriesParser.checkUpdate(query)){
-            HashMap<String, String> hashMap = new HashMap<>();
-            query = query.toLowerCase();
-            query = query.replaceAll("^\\s*update\\s+", "").replaceAll("\\s*;?\\s*$", "");
-            if(query.matches("^\\w+\\s+set\\s+(\\w\\s*=\\s*(\\'\\s*\\w\\s*\\')\\s*,\\s*)*\\s*(\\w\\s*=\\s*(\\'\\s*\\w\\s*\\')\\s*)$")){
-                int ii=0;
-                String name=null;
-                while(query!=" "){
-                    name+=query.charAt(ii);
-                    ii++;
-                }
-                if(!(filesHandler.isTableExist(name, currentDatabase.getName()))) {
+            query = query.replaceAll("(?i)^\\s*update\\s+", "").replaceAll("\\s*;?\\s*$", "");
+            if(query.toLowerCase().matches("^\\w+\\s+set\\s+(\\w+\\s*=\\s*([0-9]+|\\'\\s*\\w+\\s*\\')\\s*,\\s*)*\\s*(\\w+\\s*=\\s*([0-9]+|\\'\\s*\\w+\\s*\\')\\s*)$")){
+                String[] split = query.replaceAll("\\,", " ").replaceAll("=", " ").split("\\s+");
+                if(!filesHandler.isTableExist(split[0], currentDatabase.getName()))
                     throw new SQLException();
+                for (int i = 2; i < split.length; i+=2){
+                    if (!currentDatabase.getTable(split[0].toLowerCase()).containColumn(split[i].toLowerCase()))
+                        throw new SQLException();
                 }
-                query = query.replaceAll("^\\w+\\s+set\\s+", "");
-                query = query.replaceAll("\\s+", "");
-                query = query.replaceAll("\\'", "");
-                query = query.replaceAll("\\,", " ");
-                String[] split = query.split("\\s+");
-
-                for(int i=0;i<split.length;i++){
-                    String[] splitt = split[i].split("=");
-                    hashMap.put(splitt[0],splitt[1]);
-                }
-                currentDatabase.getTable(name).updateallcolumn(hashMap);
-                return currentDatabase.getTable(name).getIDCounter();
+                currentDatabase.getTable(split[0]).updateTable(split);
+                return currentDatabase.getTable(split[0]).getIDCounter();
             }
             else{
-                int ii=0;
-                String name=null;
-                while(query!=" "){
-                    name+=query.charAt(ii);
-                    ii++;
-                }
-                if(!(filesHandler.isTableExist(name, currentDatabase.getName()))) {
-                    throw new SQLException();
-                }
-                query = query.replaceAll("^\\w+\\s+set\\s+", "");
-                String[] split1 = query.split("where");
-                split1[0] = split1[0].replaceAll("\\s+", "");
-                split1[0] = split1[0].replaceAll("\\'", "");
-                split1[0] = split1[0].replaceAll("\\,", " ");
-                String[] split2 = query.split("\\s+");
-                for(int i=0;i<split2.length;i++){
-                    String[] splitt = split2[i].split("=");
-                    hashMap.put(splitt[0],splitt[1]);
-                }
-
-                String[] split = split1[1].split("\\s+");
-                if (split.length != 3) {
-                    split = new String[3];
-                    int j = 0;
-                    split[0] = new String();
-                    query = query.replaceAll("\\s+", " ");
-                    for (int i = 0; i < split1[1].length(); ++i) {
-                        if (query.charAt(i) == ' ') {
-                            j++;
-                            split[j] = new String();
-                        } else if (query.charAt(i) == '=' || query.charAt(i) == '<' || query.charAt(i) == '>') {
-                            j++;
-                            split[j] = String.valueOf(query.charAt(i));
-                            j++;
-                            split[j] = new String();
-                        } else {
-                            split[j] += String.valueOf(query.charAt(i));
-                        }
-                    }
-                }
-                Table table = aSwitch.meetCondition(split[1], currentDatabase.getTable(name), split[0], split[2]);
-                table.updateallcolumn(hashMap);
-
-
-
+                return 5555;
             }
         }
-
         else
             return 0;
     }
